Filmila: app shell, auth, catalog, player, DI, and tests
- Add Core layer (Supabase, repositories, IAP, HLS, network monitor)
- Wire RootView, splash, tabs, home/detail/library/search/profile
- Add CachedAsyncImage, design tokens, localization, unit tests
- Regenerate Xcode project and schemes; add Package.resolved

+++ Filmila/Models/FilmAccessEvaluator.swift
@@ -1,12 +1,11 @@
 import Foundation
 
 enum FilmAccessEvaluator: Sendable {
-    /// Viewer may watch when the film is free or a completed payment exists (checked separately).
-    static func isFree(price: Int) -> Bool {
+    static func isFree(price: Double) -> Bool {
         price == 0
     }
 
-    static func canWatch(price: Int, hasCompletedPayment: Bool) -> Bool {
+    static func canWatch(price: Double, hasCompletedPayment: Bool) -> Bool {
         isFree(price: price) || hasCompletedPayment
     }
 }
